bitrix_to_postgres/requirements.txt
apache-airflow==2.10.4
apache-airflow-providers-postgres
psycopg2-binary
Flask
Flask-Session
flask-login
flask-wtf
Werkzeug
asyncpg
sqlalchemy[asyncio]
requests
python-dotenv

fast-bitrix24
fastapi
